+++ plugins/sidepad/hooks/markdown-blocks/index.ts
@@ -4,6 +4,7 @@ export * from './flowed-text-of';
 export * from './markdown-block';
 export * from './markdown-block-at';
 export * from './markdown-blocks-of';
+export * from './parser';
 export * from './row-of-block';
 export * from './rows-of-block';
 export * from './shown-blocks-of';