+++ plugins/sidepad/hooks/tables/index.ts
@@ -1,0 +1,4 @@
+export * as default from '.';
+export * from './table';
+export * from './table-of';
+export * from './table-rows-of';